adds header to status, properly marshals all json, adds prettyprint
