+++ pkg/status/templates/checks.tmpl
@@ -1,32 +1,34 @@
 ===== Checks =====
 
-=== Running Checks ===
-{{with .RunnerStats -}}
-{{- if not .Runs}}
-  No checks have run yet
-{{end -}}
-{{- range .Checks}}
-  == {{.CheckName}} ==
-    Total Runs: {{.TotalRuns}}
-    {{if .LastError -}}
-    Error: {{lastErrorMessage .LastError}}
-    {{lastErrorTraceback .LastError -}}
-    {{- end -}}
-{{- end -}}
-{{- end}}
-
-{{- with .LoaderStats -}}
-{{if .Errors}}
-=== Loading Errors ===
-  {{ range $checkname, $errors := .Errors }}
-  == {{$checkname}} ==
-    {{ range $kind, $err := $errors -}}
-    {{- if eq $kind "Python Check Loader" -}}
-    {{$kind}}: {{ pythonLoaderError $err -}}
-    {{- else -}}
-    {{$kind}}: {{ doNotEscape $err}}
-    {{end -}}
-    {{end -}}
+  === Running Checks ===
+  {{with .RunnerStats -}}
+  {{- if not .Runs}}
+    No checks have run yet
   {{end -}}
-{{- end}}
-{{end}}
+  {{- range .Checks}}
+    {{.CheckName}}
+    --------------
+      Total Runs: {{.TotalRuns}}
+      {{if .LastError -}}
+      Error: {{lastErrorMessage .LastError}}
+      {{lastErrorTraceback .LastError -}}
+      {{- end -}}
+  {{- end -}}
+  {{- end}}
+
+  {{- with .LoaderStats -}}
+  {{if .Errors}}
+  === Loading Errors ===
+    {{ range $checkname, $errors := .Errors}}
+    {{$checkname}}
+    --------------
+      {{ range $kind, $err := $errors -}}
+      {{- if eq $kind "Python Check Loader" -}}
+      {{$kind}}: {{ pythonLoaderError $err -}}
+      {{- else -}}
+      {{$kind}}: {{ doNotEscape $err -}}
+      {{end}}
+      {{end -}}
+    {{end -}}
+  {{- end}}
+  {{end}}
