+++ pkg/status/templates/header.tmpl
@@ -1,0 +1,9 @@
+====================
+Agent (v {{.version}})
+====================
+
+Status date: 2017-06-19 22:23:13 (6s ago)
+Pid: 7916
+Platform: Linux-3.16.0-4-amd64-x86_64-with-debian-8.8
+Python Version: 2.7.13, 64bit
+Logs: <stderr>, /var/log/datadog/collector.log, syslog:/dev/log
